8 Oct

+++ app/controllers/users_controller.rb
@@ -10,19 +10,19 @@ class UsersController < ApplicationController
   def show
     begin
       @user = User.find_by_slug!(params[:id])
+
+      render 'goals/index' and return
     rescue
       go_404
     end
-    
-    render 'goals/index' and return
   end
 
   # PATCH/PUT /users/:id.:format
   def update
     pass = true
-    
+
     params[:user][:slug] = nil if user_params[:username].present? # udating username, therefore need to update slug
-    
+
     # update password, if user has no password, e.g. from Twitter, validations are bypassed, again
     if params[:user].key?(:password) && @user.encrypted_password.present?
       if params[:user][:current_password].blank? # no current password
@@ -33,18 +33,18 @@ class UsersController < ApplicationController
         pass = false # current password is not correct
       end
     end
-    
+
     if params[:user].key?(:avatar) # updating avatar
       params[:user].delete(:avatar) if params[:user][:avatar].blank?
       params[:user].delete(:remote_avatar_url) if params[:user][:remote_avatar_url].blank?
     end
-    
+
     respond_to do |format|
       if @user.update(user_params) && pass
         sign_in(current_user.reload, :bypass => true)
-        
+
         format.html { redirect_to @user, notice: 'Your account was successfully updated.' }
-        
+
         if user_params[:username].present?
           format.json { render json: { url: user_url(@user) }}
         elsif user_params[:avatar].present? || user_params[:remote_avatar_url].present?
@@ -62,7 +62,7 @@ class UsersController < ApplicationController
             @user.errors.add(:current_password, "is not correct")
           end
         end
-        
+
         format.html { render action: 'account' }
         format.json { render json: @user.errors, status: :unprocessable_entity }
       end
@@ -72,58 +72,82 @@ class UsersController < ApplicationController
   # GET/PATCH /users/:id/finish_signup
   def finish_signup
     if request.patch? && params[:user]
+      params[:user].delete(:email) if params[:user][:email].blank?
+
       if @user.update(user_params)
         if params[:user][:follow_upmit]
-          auth = @user.authorizations.where(provider: 'twitter').first
-          
-          follower = FollowUpmitService.new(auth) # Use a service to follow @upmit and delay it to avoid hold-up
+          auth = @user.authorizations.find_by(provider: session[:provider])
+
+          follower = FollowUpmitService.new(auth)
           follower.delay.follow
         end
-        
+
         sign_in(@user, :bypass => true)
         redirect_to root_path, notice: 'Your account was successfully updated.'
       end
     end
   end
-  
+
   def confirmation_sent
   end
-  
+
   #Settings
   def account
-    
+
   end
-  
+
   def profile
 
   end
-  
+
   def preference
-    if request.post? 
-      if params[:user][:disconnect].present? && %w(twitter).include?(params[:user][:disconnect]) # disconnecting twitter
+    providers = %w(twitter facebook)
+
+    if request.post?
+      if params[:user][:switch_to].present? && providers.include?(params[:user][:switch_to]) # disconnecting social networks
+        type = params[:user][:switch_to]
+        providers.delete(type)
+        auth = @user.authorizations.find_by(provider: type)
+
+        if auth.present?
+          if @user.goals.active.where(checkin_with: providers[0]).blank?
+            @user.update_attribute(:checkin_with, type)
+          else
+            flash[:alert] = "There are goals listening for check-ins from #{ providers[0].capitalize }. You can switch form #{ providers[0].capitalize } after they are all completed."
+          end
+        end
+      elsif params[:user][:disconnect].present? && providers.include?(params[:user][:disconnect]) # disconnecting social networks
         type = params[:user][:disconnect]
-        
-        auth = @user.authorizations.where(provider: type).first
-        reminder = @user.reminders.where(type: "#{ type.capitalize }Reminder").first
-        
-        if auth.present? && reminder.present?
-          Authorization.transaction do
-            auth.destroy
-            reminder.destroy
+
+        auth = @user.authorizations.find_by(provider: type)
+        reminder = @user.reminders.find_by(type: "#{ type.capitalize }Reminder")
+
+        if auth.present?
+          if @user.authorizations.size <= 1
+            flash[:alert] = "Cannot disconnect your #{ type.capitalize } account. It is the only authorization you have with us."
+
+          elsif @user.goals.active.where(checkin_with: type).size > 0
+            flash[:alert] = "There are goals listening for check-ins from #{ type.capitalize }. You can disconnect after they are all completed."
+
+          else
+            Authorization.transaction do
+              auth.destroy
+              reminder.destroy if reminder.present?
+            end
           end
         end
       elsif params[:user][:reminder].present? # updaing reminder settings
         reminder_params = params[:user][:reminder]
         type = reminder_params[:change]
-        
+
         succeed = true
-        
+
         if %w(twitter email).include?(type)
           reminder = @user.reminders.where(type: "#{ type.capitalize }Reminder").first
         else
           succeed = false
         end
-        
+
         respond_to do |format|
           if reminder.update(state: reminder_params[type]) && succeed
             format.js { head :no_content }
@@ -134,7 +158,7 @@ class UsersController < ApplicationController
       end
     end
   end
-  
+
 
   private
   def set_current_user
@@ -142,10 +166,10 @@ class UsersController < ApplicationController
   end
 
   def user_params
-    accessible = [ :first_name, :last_name, :username, :email, :timezone, :slug, :password, :password_confirmation, :bio, :avatar, :remote_avatar_url ] # extend with your own params
+    accessible = [ :display_name, :username, :email, :timezone, :slug, :avatar, :remote_avatar_url ] # extend with your own params
     params.require(:user).permit(accessible)
   end
-  
+
   def ensure_signup_complete
     # Ensure we don't go into an infinite loop
     return if action_name == 'finish_signup'
@@ -153,7 +177,7 @@ class UsersController < ApplicationController
     # Redirect to the 'finish_signup' page if the user
     # email hasn't been verified yet
     if current_user && !current_user.email_valid?
-      redirect_to finish_signup_path(current_user)
+      redirect_to finish_signup_path
     end
   end
 end
